compliance check cmd for dca
